fix run logs subscription
