+++ python_modules/dagster/dagster/_cli/workspace/cli_target.py
@@ -7,9 +7,10 @@ from typing import Any, Callable, Optional, TypeVar, cast
 
 import click
 from click import UsageError
+from dagster_shared.cli import ClickOption, apply_click_params, generate_python_pointer_options
 from dagster_shared.seven import JSONDecodeError, json
 from dagster_shared.yaml_utils import load_yaml_from_glob_list
-from typing_extensions import Never, Self, TypeAlias
+from typing_extensions import Never, Self
 
 import dagster._check as check
 from dagster import __version__ as dagster_version
@@ -416,7 +417,7 @@ class RepositoryOpts:
 
 def run_config_option(*, name: str, command_name: str) -> Callable[[T_Callable], T_Callable]:
     def wrap(f: T_Callable) -> T_Callable:
-        return _apply_click_params(f, _generate_run_config_option(name, command_name))
+        return apply_click_params(f, _generate_run_config_option(name, command_name))
 
     return wrap
 
@@ -425,39 +426,26 @@ def job_name_option(f: Optional[T_Callable] = None, *, name: str) -> T_Callable:
     if f is None:
         return lambda f: job_name_option(f, name=name)  # type: ignore
     else:
-        return _apply_click_params(f, _generate_job_name_option(name))
+        return apply_click_params(f, _generate_job_name_option(name))
 
 
 def repository_name_option(f: Optional[T_Callable] = None, *, name: str) -> T_Callable:
     if f is None:
         return lambda f: repository_name_option(f, name=name)  # type: ignore
     else:
-        return _apply_click_params(f, _generate_repository_name_option(name))
+        return apply_click_params(f, _generate_repository_name_option(name))
 
 
 def workspace_options(f: T_Callable) -> T_Callable:
-    return _apply_click_params(f, *_generate_workspace_options())
-
-
-def python_pointer_options(f: T_Callable) -> T_Callable:
-    return _apply_click_params(f, *_generate_python_pointer_options(allow_multiple=False))
+    return apply_click_params(f, *_generate_workspace_options())
 
 
 def repository_options(f: T_Callable) -> T_Callable:
-    return _apply_click_params(
+    return apply_click_params(
         f,
         _generate_repository_name_option("repository"),
         _generate_code_location_name_option("location"),
     )
-
-
-ClickOption: TypeAlias = Callable[[T_Callable], T_Callable]
-
-
-def _apply_click_params(command: T_Callable, *click_params: ClickOption) -> T_Callable:
-    for click_param in click_params:
-        command = click_param(command)
-    return command
 
 
 # ########################
@@ -523,61 +511,6 @@ def _generate_run_config_option(name: str, command_name: str) -> ClickOption:
     )
 
 
-def _generate_python_pointer_options(allow_multiple: bool) -> Sequence[ClickOption]:
-    return [
-        click.option(
-            "--working-directory",
-            "-d",
-            help="Specify working directory to use when loading the repository or job",
-            envvar="DAGSTER_WORKING_DIRECTORY",
-        ),
-        click.option(
-            "--python-file",
-            "-f",
-            # Checks that the path actually exists lower in the stack, where we
-            # are better equipped to surface errors
-            type=click.Path(exists=False),
-            multiple=allow_multiple,
-            help=(
-                "Specify python file "
-                + ("or files (flag can be used multiple times) " if allow_multiple else "")
-                + "where dagster definitions reside as top-level symbols/variables and load "
-                + ("each" if allow_multiple else "the")
-                + " file as a code location in the current python environment."
-            ),
-            envvar="DAGSTER_PYTHON_FILE",
-        ),
-        click.option(
-            "--module-name",
-            "-m",
-            multiple=allow_multiple,
-            help=(
-                "Specify module "
-                + ("or modules (flag can be used multiple times) " if allow_multiple else "")
-                + "where dagster definitions reside as top-level symbols/variables and load "
-                + ("each" if allow_multiple else "the")
-                + " module as a code location in the current python environment."
-            ),
-            envvar="DAGSTER_MODULE_NAME",
-        ),
-        click.option(
-            "--package-name",
-            multiple=allow_multiple,
-            help="Specify Python package where repository or job function lives",
-            envvar="DAGSTER_PACKAGE_NAME",
-        ),
-        click.option(
-            "--attribute",
-            "-a",
-            help=(
-                "Attribute that is either a 1) repository or job or "
-                "2) a function that returns a repository or job"
-            ),
-            envvar="DAGSTER_ATTRIBUTE",
-        ),
-    ]
-
-
 def _generate_grpc_server_options(hidden=False) -> Sequence[ClickOption]:
     return [
         click.option(
@@ -621,7 +554,7 @@ def _generate_workspace_options() -> Sequence[ClickOption]:
             type=click.Path(exists=True),
             help="Path to workspace file. Argument can be provided multiple times.",
         ),
-        *_generate_python_pointer_options(allow_multiple=True),
+        *generate_python_pointer_options(allow_multiple=True),
         *_generate_grpc_server_options(),
     ]
